requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
passlib
bcrypt
python-jose